# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-recurring",
    "version": "0.1.0",
    "description": "Standing orders for agents — mandate documents plus a client-side scheduler that pays per run and collects run reports\n\nPaid routes are settled with x402 (HTTP 402) in USDC on **Base or Solana** — the client picks the rail from the 402 `accepts` array.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "contact": {
      "name": "nirholas",
      "email": "nichxbt@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4021",
      "description": "Local development"
    },
    {
      "url": "https://{host}",
      "description": "Self-hosted",
      "variables": {
        "host": {
          "default": "example.com"
        }
      }
    }
  ],
  "tags": [
    {
      "name": "paid",
      "description": "x402-paid routes — the artifact is returned in the 200 body"
    },
    {
      "name": "free",
      "description": "Free routes"
    }
  ],
  "paths": {
    "/mandates": {
      "post": {
        "summary": "Create a signed mandate document (task, cadence, price per run, budget, expiry).",
        "description": "Create a signed mandate document (task, cadence, price per run, budget, expiry). Free.",
        "operationId": "createMandate",
        "tags": [
          "free"
        ],
        "responses": {
          "201": {
            "description": "Signed mandate document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mandateId": {
                      "type": "string"
                    },
                    "document": {
                      "const": "x402-recurring/mandate"
                    },
                    "task": {
                      "type": "object"
                    },
                    "cadence": {
                      "type": "object"
                    },
                    "pricePerRun": {
                      "type": "string"
                    },
                    "maxRuns": {
                      "type": "integer"
                    },
                    "payer": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "exhausted",
                        "expired"
                      ]
                    },
                    "runsCompleted": {
                      "type": "integer"
                    },
                    "signature": {
                      "type": "string"
                    },
                    "algorithm": {
                      "const": "HMAC-SHA256"
                    }
                  }
                },
                "example": {
                  "mandateId": "mnd_9f0c2b41-3f2a-4d1e-9a77-0b5a1c3e77aa",
                  "document": "x402-recurring/mandate",
                  "task": {
                    "type": "heartbeat",
                    "params": {
                      "label": "nightly liveness"
                    }
                  },
                  "cadence": {
                    "everySeconds": 3600
                  },
                  "pricePerRun": "$0.005",
                  "maxRuns": 10,
                  "payer": "0xAgentWalletOrSolanaPubkey",
                  "createdAt": "2026-01-01T00:00:00.000Z",
                  "expiresAt": "2026-01-31T00:00:00.000Z",
                  "status": "active",
                  "runsCompleted": 0,
                  "signature": "3f8a…hex hmac…",
                  "algorithm": "HMAC-SHA256"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "task": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "heartbeat",
                          "counter",
                          "digest"
                        ]
                      },
                      "params": {
                        "type": "object"
                      }
                    }
                  },
                  "cadence": {
                    "type": "object",
                    "properties": {
                      "everySeconds": {
                        "type": "integer"
                      }
                    }
                  },
                  "pricePerRun": {
                    "type": "string"
                  },
                  "maxRuns": {
                    "type": "integer"
                  },
                  "validDays": {
                    "type": "integer"
                  },
                  "payer": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "task": {
                  "type": "heartbeat",
                  "params": {
                    "label": "nightly liveness"
                  }
                },
                "cadence": {
                  "everySeconds": 3600
                },
                "pricePerRun": "$0.005",
                "maxRuns": 10,
                "validDays": 30,
                "payer": "0xYourAgentWallet"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Summaries of every mandate this instance has issued.",
        "description": "Summaries of every mandate this instance has issued. Free.",
        "operationId": "listMandates",
        "tags": [
          "free"
        ],
        "responses": {
          "200": {
            "description": "Mandate summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mandates": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                },
                "example": {
                  "mandates": [
                    {
                      "mandateId": "mnd_9f0c2b41-3f2a-4d1e-9a77-0b5a1c3e77aa",
                      "status": "active",
                      "pricePerRun": "$0.005",
                      "runsCompleted": 3,
                      "maxRuns": 10
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/execute/{mandateId}": {
      "post": {
        "summary": "Execute one run of a mandate and return the signed run report.",
        "description": "Execute one run of a mandate and return the signed run report. Paid: $0.005 (per mandate — its own `pricePerRun`) in USDC on Base or Solana (x402).",
        "operationId": "executeMandate",
        "tags": [
          "paid"
        ],
        "parameters": [
          {
            "name": "mandateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The `mandateId` from `POST /mandates`"
          }
        ],
        "responses": {
          "200": {
            "description": "Signed run report: run number, budget state, and the task artifact",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 settlement receipt: { rail, network, transaction, payer }",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string"
                    },
                    "document": {
                      "const": "x402-recurring/run-report"
                    },
                    "mandateId": {
                      "type": "string"
                    },
                    "runNumber": {
                      "type": "integer"
                    },
                    "executedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "charged": {
                      "type": "string"
                    },
                    "budget": {
                      "type": "object"
                    },
                    "artifact": {
                      "type": "object"
                    },
                    "signature": {
                      "type": "string"
                    },
                    "algorithm": {
                      "const": "HMAC-SHA256"
                    },
                    "payment": {
                      "type": "object",
                      "description": "Settlement receipt: rail, network, transaction, payer"
                    }
                  }
                },
                "example": {
                  "runId": "run_1a2b3c4d-5e6f-4a7b-8c9d-0e1f2a3b4c5d",
                  "document": "x402-recurring/run-report",
                  "mandateId": "mnd_9f0c2b41-3f2a-4d1e-9a77-0b5a1c3e77aa",
                  "runNumber": 3,
                  "executedAt": "2026-01-03T00:00:00.000Z",
                  "charged": "$0.005",
                  "budget": {
                    "maxRuns": 10,
                    "runsCompleted": 3,
                    "remainingRuns": 7
                  },
                  "artifact": {
                    "taskType": "heartbeat",
                    "alive": true,
                    "serverTime": "2026-01-03T00:00:00.000Z",
                    "serverUptimeSeconds": 8421,
                    "label": "nightly liveness"
                  },
                  "signature": "9c1d…hex hmac…",
                  "algorithm": "HMAC-SHA256",
                  "payment": {
                    "rail": "evm",
                    "network": "base-sepolia",
                    "transaction": "0xabc…",
                    "payer": "0xPayer…",
                    "amount": "5000"
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment required — dual-rail x402 challenge listing a Base (EVM) and a Solana entry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequired"
                }
              }
            }
          },
          "404": {
            "description": "Unknown mandateId — not charged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "All runs used — not charged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "payload": {}
                }
              },
              "example": {
                "payload": {
                  "any": "json",
                  "used_by": "the digest task"
                }
              }
            }
          }
        }
      }
    },
    "/mandates/{id}": {
      "get": {
        "summary": "Current mandate document including status and runs completed.",
        "description": "Current mandate document including status and runs completed. Free.",
        "operationId": "getMandate",
        "tags": [
          "free"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The mandateId"
          }
        ],
        "responses": {
          "200": {
            "description": "Mandate status snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "mandateId": "mnd_9f0c2b41-3f2a-4d1e-9a77-0b5a1c3e77aa",
                  "document": "x402-recurring/mandate",
                  "task": {
                    "type": "heartbeat",
                    "params": {
                      "label": "nightly liveness"
                    }
                  },
                  "cadence": {
                    "everySeconds": 3600
                  },
                  "pricePerRun": "$0.005",
                  "maxRuns": 10,
                  "payer": "0xAgentWalletOrSolanaPubkey",
                  "createdAt": "2026-01-01T00:00:00.000Z",
                  "expiresAt": "2026-01-31T00:00:00.000Z",
                  "status": "active",
                  "runsCompleted": 3,
                  "signature": "3f8a…hex hmac…",
                  "algorithm": "HMAC-SHA256"
                }
              }
            }
          },
          "404": {
            "description": "Unknown mandateId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/verify": {
      "post": {
        "summary": "Verify the HMAC-SHA256 signature of any artifact issued by this service.",
        "description": "Verify the HMAC-SHA256 signature of any artifact issued by this service. Free.",
        "operationId": "verifyArtifact",
        "tags": [
          "free"
        ],
        "responses": {
          "200": {
            "description": "`{ valid: true | false }`",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    }
                  }
                },
                "example": {
                  "valid": true
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "mandateId": "mnd_9f0c2b41-3f2a-4d1e-9a77-0b5a1c3e77aa",
                "document": "x402-recurring/mandate",
                "task": {
                  "type": "heartbeat",
                  "params": {
                    "label": "nightly liveness"
                  }
                },
                "cadence": {
                  "everySeconds": 3600
                },
                "pricePerRun": "$0.005",
                "maxRuns": 10,
                "payer": "0xAgentWalletOrSolanaPubkey",
                "createdAt": "2026-01-01T00:00:00.000Z",
                "expiresAt": "2026-01-31T00:00:00.000Z",
                "status": "active",
                "runsCompleted": 0,
                "signature": "3f8a…hex hmac…",
                "algorithm": "HMAC-SHA256"
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Liveness probe.",
        "description": "Liveness probe. Free.",
        "operationId": "health",
        "tags": [
          "free"
        ],
        "responses": {
          "200": {
            "description": "`{ ok: true }`",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "service": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "ok": true,
                  "service": "x402-recurring"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PaymentRequired": {
        "type": "object",
        "description": "x402 challenge. `accepts` always contains both an EVM and a Solana entry.",
        "required": [
          "x402Version",
          "accepts"
        ],
        "properties": {
          "x402Version": {
            "type": "integer",
            "const": 1
          },
          "error": {
            "type": "string"
          },
          "accepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentRequirements"
            }
          }
        },
        "example": {
          "x402Version": 1,
          "error": "X-PAYMENT header is required",
          "accepts": [
            {
              "scheme": "exact",
              "network": "base-sepolia",
              "maxAmountRequired": "10000",
              "resource": "http://localhost:4021/execute/:mandateId",
              "description": "Execute one run of a mandate and return the signed run report.",
              "mimeType": "application/json",
              "payTo": "0x40252CFDF8B20Ed757D61ff157719F33Ec332402",
              "maxTimeoutSeconds": 120,
              "asset": "0x036CbD53842c5426634e7929541eC2318f3dCF7e",
              "extra": {
                "name": "USDC",
                "version": "2"
              }
            },
            {
              "scheme": "exact",
              "network": "solana-devnet",
              "maxAmountRequired": "10000",
              "resource": "http://localhost:4021/execute/:mandateId",
              "description": "Execute one run of a mandate and return the signed run report.",
              "mimeType": "application/json",
              "payTo": "WwwuGbqHrwF5RG89KhUbmRWEvjnRH9k5kVM5p7T3WwW",
              "maxTimeoutSeconds": 120,
              "asset": "4zMMC9srt5Ri5X14GAgXhaHii3GnPAEERYPJgZJDncDU",
              "extra": {
                "feePayer": "CKPKJWNdJEqa81x7CkZ14BVPiY6y16Sxs7owznqtWYp5"
              }
            }
          ]
        }
      },
      "PaymentRequirements": {
        "type": "object",
        "required": [
          "scheme",
          "network",
          "maxAmountRequired",
          "resource",
          "description",
          "mimeType",
          "payTo",
          "maxTimeoutSeconds",
          "asset"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "const": "exact"
          },
          "network": {
            "type": "string",
            "enum": [
              "base-sepolia",
              "base",
              "solana-devnet",
              "solana"
            ]
          },
          "maxAmountRequired": {
            "type": "string",
            "description": "Amount in USDC base units (6 decimals)"
          },
          "resource": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "payTo": {
            "type": "string",
            "description": "EVM 0x address or Solana base58 pubkey depending on the rail"
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "asset": {
            "type": "string",
            "description": "USDC contract address or SPL mint"
          },
          "outputSchema": {
            "type": "object"
          },
          "extra": {
            "type": "object",
            "description": "EIP-712 domain on the EVM rail; `{ feePayer }` sponsor account on the Solana rail"
          }
        }
      },
      "Signature": {
        "type": "object",
        "description": "Every artifact carries an HMAC-SHA256 signature over its canonical JSON.",
        "properties": {
          "signature": {
            "type": "string",
            "description": "hex HMAC-SHA256"
          },
          "algorithm": {
            "type": "string",
            "const": "HMAC-SHA256"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}